requirements.txt
Flask==3.0.2
requests==2.31.0
azure-storage-blob
flask-cors
bs4
lxml
tzdata
flask_limiter

